Update user pattern to optional with custom response entity

+++ src/main/java/henry/goldencinema/controller/UserRestController.java
@@ -1,15 +1,17 @@
 package henry.goldencinema.controller;
 
 import henry.goldencinema.entity.User;
+import henry.goldencinema.entity.responses.ApiResponse;
 import henry.goldencinema.service.UserServices;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.security.access.prepost.PreAuthorize;
-import org.springframework.security.core.userdetails.UsernameNotFoundException;
 import org.springframework.web.bind.annotation.*;
 
-import java.net.URI;
+import java.time.LocalDateTime;
 import java.util.Collection;
+import java.util.Optional;
 
 @CrossOrigin
 @RestController
@@ -20,41 +22,89 @@ public class UserRestController {
     private UserServices userServices;
 
     @GetMapping("/all")
-    public Collection<User> getAllUsers() throws Exception {
-        return userServices.getAllUsers();
+    public ResponseEntity<?> getAllUsers() {
+        Optional<Collection<User>> users = userServices.getAllUsers();
+
+        assert users.isPresent();
+        if (users.get().isEmpty())
+            return ResponseEntity.status(HttpStatus.OK).body(new ApiResponse(
+                    HttpStatus.BAD_REQUEST.value(), LocalDateTime.now().toString(), "Empty users list", ""));
+
+        return ResponseEntity.status(HttpStatus.OK).body(new ApiResponse(
+                HttpStatus.OK.value(), LocalDateTime.now().toString(), "Users list", users));
     }
 
     @GetMapping("/{email}")
-    public User getUserByEmail(@PathVariable String email) throws UsernameNotFoundException {
-        return userServices.getUserByEmail(email);
+    public ResponseEntity<?> getUserByEmail(@PathVariable String email) {
+        Optional<User> user = userServices.getUserByEmail(email);
+
+        if (user.isPresent())
+            return ResponseEntity.status(HttpStatus.OK).body(new ApiResponse(
+                    HttpStatus.BAD_REQUEST.value(), LocalDateTime.now().toString(),
+                    "User not found for email: " + email, ""));
+
+        return ResponseEntity.status(HttpStatus.OK).body(new ApiResponse(
+                HttpStatus.OK.value(), LocalDateTime.now().toString(),
+                "User found for email: " + email, user));
     }
 
     @GetMapping("/id/{id}")
-    public User getUserById(@PathVariable String id) throws UsernameNotFoundException {
-        return userServices.getUserById(id);
+    public ResponseEntity<?> getUserById(@PathVariable String id) {
+        Optional<User> user = userServices.getUserById(id);
+
+        if (user.isPresent())
+            return ResponseEntity.status(HttpStatus.OK).body(new ApiResponse(
+                    HttpStatus.BAD_REQUEST.value(), LocalDateTime.now().toString(),
+                    "User not found for id: " + id, ""));
+
+        return ResponseEntity.status(HttpStatus.OK).body(new ApiResponse(
+                HttpStatus.OK.value(), LocalDateTime.now().toString(),
+                "User found for id: " + id, user));
     }
 
     @PostMapping
-    public ResponseEntity<User> addUser(@RequestBody User user) throws Exception {
-        User savedUser = userServices.addUser(user);
+    public ResponseEntity<?> addUser(@RequestBody User user) {
+        Optional<User> existedUser = userServices.getUserByEmail(user.getEmail());
 
-        return ResponseEntity.created(new URI("/user" + savedUser.getId())).body(savedUser);
+        if (existedUser.isPresent())
+            return ResponseEntity.status(HttpStatus.OK).body(new ApiResponse(
+                    HttpStatus.BAD_REQUEST.value(), LocalDateTime.now().toString(),
+                    "User already exist for email: " + user.getEmail(), ""));
+
+        return ResponseEntity.status(HttpStatus.OK).body(new ApiResponse(
+                HttpStatus.OK.value(), LocalDateTime.now().toString(),
+                "User created successfully", userServices.addUser(user)));
     }
 
     @PutMapping()
     @PreAuthorize("#user.email = authentication.principal.username")
-    public ResponseEntity<User> updateUser(@RequestBody User user) throws Exception {
-        User updatedUser = userServices.updateUser(user);
+    public ResponseEntity<?> updateUser(@RequestBody User user) {
+        Optional<User> existedUser = userServices.getUserById(user.getId());
 
-        return ResponseEntity.ok(updatedUser);
+        if (existedUser.isPresent())
+            return ResponseEntity.status(HttpStatus.OK).body(new ApiResponse(
+                    HttpStatus.BAD_REQUEST.value(), LocalDateTime.now().toString(),
+                    "User not found", user));
+
+        return ResponseEntity.status(HttpStatus.OK).body(new ApiResponse(
+                HttpStatus.OK.value(), LocalDateTime.now().toString(),
+                "User created successfully", userServices.updateUser(user)));
     }
 
     @DeleteMapping("/{id}")
-    @PreAuthorize("#id = userServicesImpl.getUserByEmail(authentication.principal.username).id")
-    public ResponseEntity<User> deleteUser(@PathVariable String id) throws Exception {
-        userServices.deleteUserById(id);
+    @PreAuthorize("#id = userServicesImpl.getUserByEmail(authentication.principal.username).get().id")
+    public ResponseEntity<?> deleteUser(@PathVariable String id) {
+        Optional<User> existedUser = userServices.getUserById(id);
 
-        return ResponseEntity.ok().build();
+        if (existedUser.isPresent())
+            return ResponseEntity.status(HttpStatus.OK).body(new ApiResponse(
+                    HttpStatus.BAD_REQUEST.value(), LocalDateTime.now().toString(),
+                    "User not found for id: " + id, ""));
+
+        userServices.deleteUserById(id);
+        return ResponseEntity.status(HttpStatus.OK).body(new ApiResponse(
+                HttpStatus.OK.value(), LocalDateTime.now().toString(),
+                "User deleted successfully", ""));
     }
 
 }
